# backend/ticket-service/docs/swagger.yaml
basePath: /
definitions:
...
  domain.TicketAttachment:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      size:
        type: integer
      ticket_id:
        type: integer
      uploader_id:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Tickets, comments, attachments and reports as clients see them: through the API Gateway.
    "Try it out" calls the gateway, so the origin of this page must be listed in the gateway ALLOWED_ORIGINS.
    Send the access token from POST /auth/login as `Bearer <token>` in the Authorization header.
  title: Helpdesk Ticket API
  version: "1.0"
paths:
...
  /tickets/{id}/attachments:
    get:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.TicketAttachment'
                  type: array
              type: object
        "404":
          description: ticket not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List ticket attachments (metadata only)
      tags:
      - Attachments
    post:
      consumes:
      - multipart/form-data
      description: One file per request, 5 MB at most. Rate limited.
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: File to attach
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.TicketAttachment'
              type: object
        "400":
          description: missing file
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: ticket not found
          schema:
            $ref: '#/definitions/response.Response'
        "413":
          description: attachment too large (max 5MB)
          schema:
            $ref: '#/definitions/response.Response'
        "429":
          description: rate limited
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Upload an attachment
      tags:
      - Attachments
  /tickets/{id}/attachments/{attachmentId}:
    get:
      description: Returns the file content with its original Content-Type.
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attachment ID
        in: path
        name: attachmentId
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: attachment not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Download an attachment
      tags:
      - Attachments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"